Tentato di fare il modifica vacanza. Problema con l'aggiornamento della view prima che la vacanza sia caricata (sempre vuota). Prova ad usare LiveData su vacanza e partecipanti, magari tutto in altra activity. - in recenti, aggiungi un header e preleva le 3 liste a seconda dell'ordine temporale - in archivio, visualizza vacanze archiviate - modifica una vacanza - rimuovi tastiera quando si seleziona la data e in generale quando non serve al campo successivo - validazione dei campi della nuova vacanza

+++ app/src/main/java/com/rbiffi/vacationfriend/Repository/DAOs/IVacationDao.java
@@ -18,7 +18,7 @@ public interface IVacationDao {
 
 //READ
     @Query("SELECT * FROM Vacation WHERE id = :id")
-    Vacation getVacationDetails(int id);
+    Vacation getVacationDetails(long id);
 
     @Query("SELECT id, title, startDate, endDate, photo " +
             "FROM Vacation " +
